91 - Moves non-generic objects and enums out of WFOFilterTabs into the consuming components
