+++ apps/wfo-ui/pages/subscriptions/index.tsx
@@ -4,13 +4,11 @@ import { StringParam, useQueryParam, withDefault } from 'use-query-params';
 import { START_DATE } from '../../components/Subscriptions/subscriptionsQuery';
 import {
     DEFAULT_PAGE_SIZE,
-    defaultSubscriptionsTabs,
     getSortDirectionFromString,
     getSubscriptionsTabTypeFromString,
     getTableConfigFromLocalStorage,
     SortOrder,
     WFOFilterTabs,
-    WFOSubscriptionsTabType,
     useDataDisplayParams,
 } from '@orchestrator-ui/orchestrator-ui-components';
 import { EuiPageHeader, EuiSpacer } from '@elastic/eui';
@@ -20,6 +18,55 @@ import {
 } from '../../components/Subscriptions/Subscriptions';
 import NoSSR from 'react-no-ssr';
 import { SUBSCRIPTIONS_TABLE_LOCAL_STORAGE_KEY } from '../../constants';
+import { WFOFilterTab } from '@orchestrator-ui/orchestrator-ui-components/src';
+
+// Todo: consider to move this out of this component
+export enum WFOSubscriptionsTabType {
+    ACTIVE = 'ACTIVE',
+    TERMINATED = 'TERMINATED',
+    TRANSIENT = 'TRANSIENT',
+    ALL = 'ALL',
+}
+
+export const defaultSubscriptionsTabs: WFOFilterTab<
+    WFOSubscriptionsTabType,
+    Subscription
+>[] = [
+    {
+        id: WFOSubscriptionsTabType.ACTIVE,
+        translationKey: 'active',
+        alwaysOnFilters: [
+            {
+                field: 'status',
+                value: 'active',
+            },
+        ],
+    },
+    {
+        id: WFOSubscriptionsTabType.TERMINATED,
+        translationKey: 'terminated',
+        alwaysOnFilters: [
+            {
+                field: 'status',
+                value: 'terminated',
+            },
+        ],
+    },
+    {
+        id: WFOSubscriptionsTabType.TRANSIENT,
+        translationKey: 'transient',
+        alwaysOnFilters: [
+            {
+                field: 'status',
+                value: 'initial-provisioning-migrating',
+            },
+        ],
+    },
+    {
+        id: WFOSubscriptionsTabType.ALL,
+        translationKey: 'all',
+    },
+];
 
 export default function SubscriptionsPage() {
     const router = useRouter();
