+++ packages/orchestrator-ui-components/src/components/WFOSubscriptionsTabs/WFOFilterTabs.tsx
@@ -2,103 +2,29 @@ import React from 'react';
 import { EuiTab, EuiTabs } from '@elastic/eui';
 import { useTranslations } from 'next-intl';
 
-export enum WFOSubscriptionsTabType {
-    ACTIVE = 'ACTIVE',
-    TERMINATED = 'TERMINATED',
-    TRANSIENT = 'TRANSIENT',
-    ALL = 'ALL',
-}
-
-export enum WFOProcessListTabType {
-    ACTIVE = 'ACTIVE',
-    COMPLETED = 'COMPLETED',
-}
-
-export const defaultSubscriptionsTabs: WFOFilterTab<WFOSubscriptionsTabType>[] =
-    [
-        {
-            id: WFOSubscriptionsTabType.ACTIVE,
-            translationKey: 'active',
-            alwaysOnFilters: [
-                {
-                    field: 'status',
-                    value: 'active',
-                },
-            ],
-        },
-        {
-            id: WFOSubscriptionsTabType.TERMINATED,
-            translationKey: 'terminated',
-            alwaysOnFilters: [
-                {
-                    field: 'status',
-                    value: 'terminated',
-                },
-            ],
-        },
-        {
-            id: WFOSubscriptionsTabType.TRANSIENT,
-            translationKey: 'transient',
-            alwaysOnFilters: [
-                {
-                    field: 'status',
-                    value: 'initial-provisioning-migrating',
-                },
-            ],
-        },
-        {
-            id: WFOSubscriptionsTabType.ALL,
-            translationKey: 'all',
-        },
-    ];
-
-export const defaultProcessListTabs: WFOFilterTab<WFOProcessListTabType>[] = [
-    {
-        id: WFOProcessListTabType.ACTIVE,
-        translationKey: 'active',
-        alwaysOnFilters: [
-            {
-                field: 'status',
-                value: 'created-running-suspended-waiting-failed-resumed',
-            },
-        ],
-    },
-    {
-        id: WFOProcessListTabType.COMPLETED,
-        translationKey: 'completed',
-        alwaysOnFilters: [
-            {
-                field: 'status',
-                value: 'active',
-            },
-        ],
-    },
-];
-
-////////
-
-export type FilterQuery = {
-    field: string;
+export type FilterQuery<DataType> = {
+    field: keyof DataType;
     value: string;
 };
 
-export type WFOFilterTab<T> = {
-    id: T;
+export type WFOFilterTab<TabType, DataType> = {
+    id: TabType;
     translationKey: string;
-    alwaysOnFilters?: FilterQuery[];
+    alwaysOnFilters?: FilterQuery<DataType>[];
 };
 
-export type WFOFilterTabsProps<T> = {
-    tabs: WFOFilterTab<T>[];
-    selectedSubscriptionsTab: T;
-    onChangeSubscriptionsTab: (updatedSubscriptionsTab: T) => void;
+export type WFOFilterTabsProps<TabType, DataType> = {
+    tabs: WFOFilterTab<TabType, DataType>[];
+    selectedSubscriptionsTab: TabType;
+    onChangeSubscriptionsTab: (updatedSubscriptionsTab: TabType) => void;
 };
 
-export const WFOFilterTabs = <T extends string>({
+export const WFOFilterTabs = <TabType extends string, DataType>({
     tabs,
     selectedSubscriptionsTab,
     onChangeSubscriptionsTab,
-}: WFOFilterTabsProps<T>) => {
+}: WFOFilterTabsProps<TabType, DataType>) => {
+    // Todo, introduce a prop to pass the translation key
     const t = useTranslations('subscriptions.tabs');
     return (
         <EuiTabs>
